api/requirements.txt
clip-by-openai==1.1
colorama==0.4.6
chromadb==0.5.5
faiss-cpu==1.8.0.post1
filelock==3.15.4
fsspec==2024.6.1
ftfy==6.2.3
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pillow==10.4.0
sympy==1.13.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.66.5
typing-extensions==4.12.2
wcwidth==0.2.13
pandas